feat: Introduce Add Person screen and refactor navigation
This commit introduces a new "Add Person" screen and refactors several parts of the application, including data models, navigation, and event handling.

-   **New "Add Person" Screen:**
    -   A new `AddPersonPage` is created, allowing users to add a new person with a first name, last name, and address.
    -   `AddPersonViewModel` and `AddEvent` are introduced to manage state and user actions for this new screen.
    -   Reverse geocoding is implemented in the `AddPersonViewModel` to automatically fetch an address when adding a person from a map location.
    -   Navigation is set up to allow adding a person from the list (via a new FAB) or directly from a long-press on the map, passing coordinates to pre-fill the address.

-   **Data Model & Database Changes:**
    -   The `id` field in `Person`, `PersonEntity`, `PersonDao`, and `PersonRepository` has been changed from `Int` to `Long`. This affects all related functions and navigation arguments.

-   **Map Screen Refactoring:**
    -   Removed the logic for creating a person directly from the `MapsViewModel` upon a long-press.
    -   The `onMapLongClick` event now navigates to the new `AddPersonPage`, passing the selected `LatLng` coordinates.
    -   Deleted the now-unused `MapEvent.kt` file.

-   **UI & UX Improvements:**
    -   **List Screen:**
        -   A Floating Action Button (FAB) has been added to navigate to the `AddPersonPage`.
        -   The display of address and coordinates in the list item has been adjusted for clarity.
    -   **Update Screen:**
        -   A delete button with an icon has been added to the top bar, allowing for the deletion of a person.
    -   **General:**
        -   Added new string resources for UI text and internationalization (French).
        -   The placeholder `Detail.kt` file has been removed.
        -   Navigation logic in `MainActivity` is updated to accommodate the new "Add Person" flow and the data type change for `personId`.

+++ app/src/main/java/com/lucwaw/friendsLocal/data/PersonDao.kt
@@ -11,7 +11,7 @@ import kotlinx.coroutines.flow.Flow
 interface PersonDao {
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertPerson(spot: PersonEntity): Int
+    suspend fun insertPerson(spot: PersonEntity): Long
 
     @Delete
     suspend fun deletePerson(spot: PersonEntity)
@@ -20,5 +20,5 @@ interface PersonDao {
     fun getPersons(): Flow<List<PersonEntity>>
 
     @Query("SELECT * FROM personentity WHERE id = :id")
-    suspend fun getPersonById(id: Int): PersonEntity?
+    suspend fun getPersonById(id: Long): PersonEntity?
 }
